Fixed compilation on 64 bit systems.

+++ source/s_flac.cpp
@@ -92,7 +92,7 @@ public:
 	protected:
 		//	FLAC decoder callbacks.
 		::FLAC__StreamDecoderReadStatus read_callback(FLAC__byte buffer[],
-			unsigned *bytes);
+			size_t* bytes);
 		::FLAC__StreamDecoderWriteStatus write_callback(
 			const ::FLAC__Frame *frame, const FLAC__int32 * const buffer[]);
 		void metadata_callback(const ::FLAC__StreamMetadata *metadata);
@@ -477,7 +477,7 @@ void VFlacAudioCodec::FStream::StrmWrite(const FLAC__int32* const Buf[],
 //==========================================================================
 
 ::FLAC__StreamDecoderReadStatus VFlacAudioCodec::FStream::read_callback(
-	FLAC__byte buffer[], unsigned* bytes)
+	FLAC__byte buffer[], size_t* bytes)
 {
 	guard(VFlacAudioCodec::FStream::read_callback);
 	if (*bytes > 0)
